refactor: move FileHandler class out of gui package
